2do avance de CMS - restricciones de noticias

+++ backend/controllers/NoticiaController.php
@@ -116,15 +116,28 @@ class NoticiaController extends Controller
      */
     public function actionUpdate($id)
     {
+        
         $model = $this->findModel($id);
-
+        $noticia=Noticia::findOne(['id'=>$id]);
+        $creado_por=\yii\helpers\ArrayHelper::getValue($noticia, 'created_by');   
+        //print_r(\yii\helpers\ArrayHelper::getValue($noticia, 'created_by'));die;
+        //Yii::$app->user->identity
+        
+        if ($creado_por == Yii::$app->user->getId()){
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
             return $this->redirect(['view', 'id' => $model->id]);
         } else {
             return $this->render('update', [
                 'model' => $model,
             ]);
+        }            
+        }else {
+            \Yii::$app->getSession()->setFlash('info', 'No tiene permisos para editar noticias que no ha publicado');
+            $this->redirect(['noticia/index']);
+            
         }
+
+
     }
 
     /**
@@ -135,9 +148,19 @@ class NoticiaController extends Controller
      */
     public function actionDelete($id)
     {
+        
+         $noticia=Noticia::findOne(['id'=>$id]);
+        $creado_por=\yii\helpers\ArrayHelper::getValue($noticia, 'created_by');
+        
+       if ($creado_por == Yii::$app->user->getId()){
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
+       }else{
+           \Yii::$app->getSession()->setFlash('info', 'No tiene permisos para borrar noticias que no ha publicado');
+            $this->redirect(['noticia/index']); 
+           
+       }
     }
 
     /**
